+++ api/core/schemas/responses.py
@@ -239,7 +239,7 @@ class ProductVersionSchema(Schema):
     version: str = Field(
         ...,
         description="Deployed package version",
-        examples=["0.8.21"],
+        examples=["0.8.22"],
     )
     license: str = Field(
         ...,